[sys-kernel/linux-sabayon-sources-2.6.32] Updated ebuild for compatibility.
Updated sys-kernel/linux-sabayon-sources-2.6.32 to provide compatiblity
with the sys-kernel/linux-sabayon package, and provided an additional
USE flag (sources_standalone) to toggle included files to provide a
working set of kernel sources that doesn't have file collisions with the
sys-kernel/linux-sabayon package.  The default for the
"sources_standalone" USE flage is disabled, which will provide a
dependency on sys-kernel/linux-sabayon-2.6.32.  This is the desired
behavior for an entropy build of this package.

+++ sys-kernel/linux-sabayon-sources/linux-sabayon-sources-2.6.32.ebuild
@@ -10,6 +10,11 @@ inherit sabayon-kernel
 KEYWORDS="~amd64 ~x86"
 DESCRIPTION="Official Sabayon Linux Standard kernel sources"
 RESTRICT="mirror"
+IUSE="sources_standalone"
+
+DEPEND="${DEPEND}
+	sources_standalone? ( !=sys-kernel/linux-sabayon-${PVR} )
+	!sources_standalone? ( =sys-kernel/linux-sabayon-${PVR} )"
 
 src_compile() {
 	kernel-2_src_compile
@@ -31,9 +36,12 @@ src_install() {
 	local oldarch=${ARCH}
 	cp ${FILESDIR}/${P}-${ARCH}.config .config || die "cannot copy kernel config"
 	unset ARCH
-	make modules_prepare || die "failed to run modules_prepare"
-	rm .config || die "cannot remove .config"
-	rm Makefile || die "cannot remove Makefile"
+	if ! use sources_standalone; then
+		make modules_prepare || die "failed to run modules_prepare"
+		rm .config || die "cannot remove .config"
+		rm Makefile || die "cannot remove Makefile"
+		rm include/linux/version.h || die "cannot remove include/linux/version.h"
+	fi
 	ARCH=${oldarch}
 
 }
